requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
psycopg[binary]>=3.2,<3.3
dj-database-url>=2.3,<2.4
django-environ>=0.11,<0.12
django-cors-headers>=4.6,<4.7
django-modeltranslation>=0.19,<0.20
resend>=2.0,<3.0
gunicorn>=23,<24
whitenoise>=6.8,<6.9
